% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{load_cov_pbulk}
\alias{load_cov_pbulk}
\title{Load COVID-19 PBMC pseudo-bulk benchmark datasets}
\usage{
load_cov_pbulk(which = "all")
}
\arguments{
\item{which}{Which pseudo-bulk dataset to load. Use \code{"all"} to load all
ten datasets, or an integer from \code{1} to \code{10} to load a specific
file. Defaults to \code{"all"}.}
}
\value{
If \code{which = "all"}, a named list of length 10. Each element is a
pseudo-bulk dataset containing:
\itemize{
\item \code{bulk_expression_profiles}: pseudo-bulk expression matrix.
\item \code{ground_truth_proportions}: true cell-type proportions for the
pseudo-bulk samples.
}

If \code{which} is an integer from \code{1} to \code{10}, only the selected
pseudo-bulk dataset is returned.
}
\description{
Loads pseudo-bulk expression profiles generated from the COVID-19 PBMC
immune test split. These pseudo-bulk datasets are intended for deconvolution
benchmarks using the corresponding \code{train_cov} reference.
}
\details{
Each packaged file contains 1000 pseudo-bulk samples and their matching
ground-truth cell-type proportions.

The files are stored in \code{inst/extdata} as
\code{cov_pbulk_1.rds}, \code{cov_pbulk_2.rds}, ...,
\code{cov_pbulk_10.rds}. They were created from \code{test_cov} and are
designed to be used with \code{train_cov} for deconvolution benchmarking.
}
\examples{
\dontrun{
pbulks <- load_cov_pbulk()

pbulk1 <- load_cov_pbulk(1)
expr <- pbulk1$bulk_expression_profiles
prop <- pbulk1$ground_truth_proportions
}

}
